Marvell-armhf: Reduce kernel image size for u-boot
Marvell-armhf platforms using u-boot has constraints when loading the
kernel image into the physical memory region reserved for the kernel
and ramdisk image.  The size constraint is platform specific.
This commit reduces the size of the kernel image to
meet those constraints. The compression used for the kernel is xz.

- Disable unused architectures to reduce the size of the kernel
- Disable sound device driver module

Testing:
 - Check kernel boots without any memory overlapping errors
 - kernel binary size before this commit : 6197760 bytes
 - kernel binary size after this commit  : 3600624 bytes

Signed-off-by: Antony Rheneus <arheneus@marvell.com>