Upgrade kernel from 4.19.152 to 5.10.40, and port some patches
As part of the Bullseye upgrade, upgrade to the 5.10.40 kernel. At this
point, 28 patches apply (some of them required some changes) onto the
kernel. Others require more changes. All patches, however, need to be
refreshed, so that the line numbers match the current version of source
code.

Also update our build script to commit changes to *.maintscript and
*.bug-presubj files, these are generated files and get changed because
of our abiname change.

Signed-off-by: Saikrishna Arcot <sarcot@microsoft.com>