[Mellanox] Remove incorrectly generated patch allegedly avoiding kernel deadlock but causes one (#231)
Patch 0028-thermal-Fix-deadlock-in-thermal-thermal_zone_device_.patch was incorrectly generated. It is supposed to backport Linux commit torvalds/linux@163b00c, but the commit was already backported to the stable series in v4.19.89, so included in Linux v.19.152 currently used by SONiC. Instead, the incorrect patch reverts the first hunk of the commit. So, remove it.

For the old patch, guess there was some mistake has been made while generating the patch. Instead of applying the fix to the code, it's generated in a opposite way, the content of the wrong patch is reverting the first part of the patch, so it can still building but make the fix incomplete.

Regression tests have been performed against these patches on the Mellanox platform, no issue found.