Add patch by Mike Manning to fix behavior of sockets bound to an interface
This patch restores the old behavior of preferring bound sockets over
unbound sockets when they are both opened by the same application, for
UDP sockets only.

The primary commit that started this issue is
torvalds/linux@6da5b0f027a825df2aebc1927a27bda185dc03d4 which prefers an
unbound datagram socket over a bound datagram socket to be chosen if the
sockets aren't in a VRF.

This breaks some assumptions that dhcrelay (from isc-dhcp) makes when it
opens a bound socket for each interface and an unbound "fallback"
socket, and it expects that the bound sockets will be preferred over the
unbound sockets.

This is committed upstream at
torvalds/linux@8d6c414cd2fb74aa6812e9bfec6178f8246c4f3a, and should be
in 5.15 and newer.

Signed-off-by: Saikrishna Arcot <sarcot@microsoft.com>