Bookworm:AMD-Pensando ELBA SOC support (#353)
This is a change request for updating the kernel patches for AMD-Pensando ELBA SOC to linux 6.1 based bookworm.

The patches has been refactored in line with the latest kernel patches released by pensando linux team.

Elba provides a secure, controlled portal to network services, storage, and the data center control plane. This SOC is used in AMD-Pensando PCI Distributed Services Card (DSC).

The changes are verified on Pensando DSC2-200 card.